Added the hexagon struct to CategoryPanel file and made a custom initializer for easier use.
